Remove broken link in android docs
This link to code signing is broken because code signing is not a top level item, it is a step under install. The ios docs also have a section on code signing and does not have a link at the top so I thought that was probably the intended behavior.